requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
langchain-google-vertexai>=3.2.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
apscheduler>=3.10.0
pytz>=2024.1
aiohttp>=3.9.0
pyyaml>=6.0
duckduckgo-search>=6.0.0
chromadb>=0.5.0
